requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
flask-login==0.6.3
python-dotenv==1.0.0
requests==2.31.0
waitress==3.0.0
psycopg2-binary==2.9.9
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.0.7
jinja2==3.1.2
